foundation::containers::spmc_queue
Struct SpmcStealQueue

pub struct SpmcStealQueue<T: Send> { /* private fields */ }

Lock free queue for Single-Producer-Multiple-Consumer, supporting stealing API and fixed size allocated during construction

Key safety assumptions:

push can happen only from single thread, pop can happen from same thread as push

stealing can happen from any thread

stealing have same order of extracting values as ‘pop’

T cannot have self referential data directly inside itself

Implementations

impl<T: Send> SpmcStealQueue<T>

pub fn new(size: u32) -> Self

Construct new queue of size size. It will panic if construction can not happen.

Arguments

size - must be power of two

pub fn capacity(&self) -> u32

Returns number of elements the queue can hold at max

pub fn count(&self) -> u32

Returns number of elements in the queue

pub fn get_local(&self) -> Option<LocalProducerConsumer<'_, T>>

Created LocalProducerConsumer if there is any existing, otherwise None

pub fn get_boundedl(self: &Arc<Self>) -> Option<BoundProducerConsumer<T>>

Created LocalProducerConsumer if there is any existing, otherwise None

pub fn steal_into(&self, dst: &Self, divider: Option<u8>) -> Option<u32>

Let caller steal a work from other queue into its queue

Arguments

dst - destination queue

divider - the call will try to steal min(half count of a queue, avail space in dst) by default. When divider is provided, it will calculate steal size as half_count / divider

pub fn steal_internal( &self, dst: &Self, dst_head: u32, max_steal_size: u32, divider: u8, ) -> u32