foundation

Module prelude

Re-exports

pub use tracing_subscriber;

Modules

span

Spans represent periods of time in which a program was executing in a particular context.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

info

Constructs an event at the info level.

span

Constructs a new span.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

ConditionVariable

Condition variable which requires a fixed predicate on creation which is then used in ConditionVariable::blocking_wait_while() and ConditionVariable::timed_wait_while() concurrently with the benefit of triggering single waiters. The reason is when one waits on multiple conditions but triggers only one waiter it is possible that a waiter is triggered which condition is not satisfied while another waiter is not signaled which could have continued. This could lead to some deadlocks.

ConditionVariableBuilder

Builder for the MultiConditionVariable and the ConditionVariable. It creates a condition variable which already includes a mutex. The design idea originated from the fact that waiting on the same condition variable with two distinct mutexes is undefined behavior according to the posix standard. An integrated mutex avoids this possible pitfall.

ConditionVariableData

Provides access to the underlying type and predicate of a ConditionVariable or with a single static condition.

ConditionVariableGuard

Is returned by the ConditionVariable in ConditionVariable::notify_all() or ConditionVariable::notify_one(). It enables the user to change the guarded content before the notification is send. The trigger is signaled when the guard goes out of scope.

FixedSizeVec

Relocatable vector with compile time fixed size capacity. In contrast to its counterpart the RelocatableVec it is movable.

IoxAtomic

iceoryx2 implementation of an atomic that has an internal [RwLockWriterPreference]. It enables atomic operations on platforms that do not support them with the restriction that those operations are no longer lock-free.

Level

Describes the level of verbosity of a span or event.

MultiConditionVariable

Condition variable which allows to use multiple conditions in MultiConditionVariable::blocking_wait_while() and MultiConditionVariable::timed_wait_while() concurrently but with the draw back that only all waiters can be triggered and not one. The reason is when one waits on multiple conditions but triggers only one waiter it is possible that a waiter is triggered which condition is not satisfied while another waiter is not signaled which could have continued. This could lead to some deadlocks.

MultiConditionVariableGuard

Is returned by the MultiConditionVariable in MultiConditionVariable::notify_all(). It enables the user to change the guarded content before the notification is send. The trigger is signaled when the guard goes out of scope.

Mutex

Represents a POSIX mutex which can be created by the MutexBuilder.

MutexBuilder

Creates a Mutex.

MutexGuard

A guard which allows the modification of a value guarded by a Mutex. It is returned in Mutex::lock(), Mutex::try_lock() and Mutex::timed_lock().

MutexHandle

Predicate

Enums

CommonErrors

ConditionVariableCreationError

ConditionVariableError

The ConditionVariableError enum is a generalization when one doesn’t require the fine-grained error handling enums. One can forward forward ConditionVariableError as more generic return value when a method returns a Condition***Error. On a higher level it is again convertable into [crate::Error].

ConditionVariableTimedWaitError

ConditionVariableWaitError

ConditionVariableWithSegregatedMutexCreationError

MutexCreationError

MutexError

The MutexError enum is a generalization when one doesn’t require the fine-grained error handling enums. One can forward MutexError as more generic return value when a method returns a Mutex***Error. On a higher level it is again convertable to [crate::Error].

MutexGetPrioCeilingError

MutexLockError

MutexPriorityInheritance

Defines how the priority of a mutex owning thread changes when another thread with an higher priority would like to acquire the mutex.

MutexSetPrioCeilingError

MutexThreadTerminationBehavior

Defines the behavior when a mutex owning thread is terminated

MutexTimedLockError

MutexType

The type of a mutex defines its behavior.

MutexUnlockError

Traits

BasicConditionVariableInterface

Handle

Represents a handle that is in general inter-process capable.

IpcCapable

Represents struct that can be configured for inter-process use.

Type Aliases

IoxAtomicBool

Behaves like core::sync::atomic::AtomicBool

IoxAtomicI8

Behaves like core::sync::atomic::AtomicI8

IoxAtomicI16

Behaves like core::sync::atomic::AtomicI16

IoxAtomicI32

Behaves like core::sync::atomic::AtomicI32

IoxAtomicI64

Behaves like core::sync::atomic::AtomicI64

IoxAtomicIsize

Behaves like core::sync::atomic::AtomicIsize

IoxAtomicU8

Behaves like core::sync::atomic::AtomicU8

IoxAtomicU16

Behaves like core::sync::atomic::AtomicU16

IoxAtomicU32

Behaves like core::sync::atomic::AtomicU32

IoxAtomicU64

Behaves like core::sync::atomic::AtomicU64

IoxAtomicUsize

Behaves like core::sync::atomic::AtomicUsize

RelocatableVec

Non-movable relocatable vector with runtime fixed size capacity.

Vec

Vector with run-time fixed size capacity. In contrast to its counterpart the RelocatableVec it is movable but is not shared memory compatible.