foundation::prelude

Struct ConditionVariableBuilder

pub struct ConditionVariableBuilder { /* private fields */ }

Builder for the MultiConditionVariable and the ConditionVariable. It creates a condition variable which already includes a mutex. The design idea originated from the fact that waiting on the same condition variable with two distinct mutexes is undefined behavior according to the posix standard. An integrated mutex avoids this possible pitfall.

Defaults

As default the condition variable is is_interprocess_capable. For the internal mutex the defaults from MutexBuilder are used.

Example

use iceoryx2_bb_posix::condition_variable::*;
use iceoryx2_bb_posix::clock::*;
use iceoryx2_bb_posix::mutex::*;

// create a condition variable which allows multiple predicates in blocking_wait_while nad
// timed_wait_while
let mtx_handle = MutexHandle::<i32>::new();
let condvar = ConditionVariableBuilder::new()
                    .is_interprocess_capable(true)
                    // this is applied to the mutex and the condition variable
                    .clock_type(ClockType::Monotonic)
                    // begin: mutex settings
                    .mutex_type(MutexType::WithDeadlockDetection)
                    .thread_termination_behavior(MutexThreadTerminationBehavior::ReleaseWhenLocked)
                    .priority_inheritance(MutexPriorityInheritance::None)
                    // end: mutex settings
                    .create_multi_condition_variable(123, &mtx_handle);

// create a condition variable which requires a predicate on creation but enables methods like
// notify_one, trigger_one ...
let mtx_handle = MutexHandle::<ConditionVariableData<i32>>::new();
let condvar = ConditionVariableBuilder::new()
                    // do not configure anything and let's use the default settings
                    .create_condition_variable(0, |v| *v >= 100, &mtx_handle);

Implementations

impl ConditionVariableBuilder

pub fn new() -> ConditionVariableBuilder

pub fn is_interprocess_capable(self, value: bool) -> ConditionVariableBuilder

Enable interprocess capabilities for the condition variable and the underlying mutex

pub fn clock_type(self, value: ClockType) -> ConditionVariableBuilder

Defines the type of clock which should be used for timed_wait and timed_wait_while.

pub fn mutex_type(self, value: MutexType) -> ConditionVariableBuilder

Defines the mutex type of the underlying mutex

pub fn thread_termination_behavior( self, value: MutexThreadTerminationBehavior, ) -> ConditionVariableBuilder

Defines the thread termination behavior of the underlying mutex

pub fn priority_inheritance( self, value: MutexPriorityInheritance, ) -> ConditionVariableBuilder

Defines the priority inheritance of the underlying mutex

pub fn priority_ceiling(self, value: Option<i32>) -> ConditionVariableBuilder

Defines the priority ceiling of the underlying mutex

pub fn create_multi_condition_variable<T>( self, t: T, handle: &MutexHandle<T>, ) -> Result<MultiConditionVariable<'_, T>, ConditionVariableCreationError>
where T: Debug,

Creates a MultiConditionVariable encapsulating the value of type T which can be modified and used for triggering.

The condition variable can use multiple conditions in MultiConditionVariable::blocking_wait_while() and MultiConditionVariable::timed_wait_while() but is only able to trigger all waiters.

pub fn create_condition_variable<T, F>( self, t: T, predicate: F, handle: &MutexHandle<ConditionVariableData<T>>, ) -> Result<ConditionVariable<'_, T>, ConditionVariableCreationError>
where T: Debug, F: Fn(&T) -> bool + 'static,

Creates a ConditionVariable encapsulating the value of type T which can be modified and used for triggering.

The condition variable has one fixed condition which has to be provided on construction. The methods ConditionVariable::blocking_wait_while() and ConditionVariable::timed_wait_while() will wait on that preset condition until it is satisfied. The restriction to a preset fixed condition comes with the feature to signal single waiters with ConditionVariable::trigger_one() for instance.