foundation::prelude
Struct IoxAtomic

#[repr(C)]
pub struct IoxAtomic<T>
where
    T: AtomicInteger,
{ /* private fields */ }

iceoryx2 implementation of an atomic that has an internal [RwLockWriterPreference]. It enables atomic operations on platforms that do not support them with the restriction that those operations are no longer lock-free.

Implementations

impl<T> IoxAtomic<T>
where T: AtomicInteger,

pub const fn new(v: T) -> IoxAtomic<T>

See core::sync::atomic::AtomicU64::new()

pub const fn as_ptr(&self) -> *mut T

See core::sync::atomic::AtomicU64::as_ptr()

pub fn compare_exchange( &self, current: T, new: T, _success: Ordering, _failure: Ordering, ) -> Result<T, T>

See core::sync::atomic::AtomicU64::compare_exchange()

pub fn compare_exchange_weak( &self, current: T, new: T, success: Ordering, failure: Ordering, ) -> Result<T, T>

See core::sync::atomic::AtomicU64::compare_exchange_weak()

pub fn fetch_add(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_add()

pub fn fetch_and(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_and()

pub fn fetch_max(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_max()

pub fn fetch_min(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_min()

pub fn fetch_nand(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_nand()

pub fn fetch_or(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_or()

pub fn fetch_sub(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_sub()

pub fn fetch_update<F>( &self, _set_order: Ordering, _fetch_order: Ordering, f: F, ) -> Result<T, T>
where F: FnMut(T) -> Option<T>,

See core::sync::atomic::AtomicU64::fetch_update()

pub fn fetch_xor(&self, value: T, order: Ordering) -> T

See core::sync::atomic::AtomicU64::fetch_xor()

pub fn into_inner(self) -> T

See core::sync::atomic::AtomicU64::into_inner()

pub fn load(&self, _order: Ordering) -> T

See core::sync::atomic::AtomicU64::load()

pub fn store(&self, value: T, _order: Ordering)

See core::sync::atomic::AtomicU64::store()

pub fn swap(&self, value: T, _order: Ordering) -> T

See core::sync::atomic::AtomicU64::swap()