foundation::prelude

Struct MultiConditionVariable

pub struct MultiConditionVariable<'mtx_handle, T>
where
    T: Debug,
{ /* private fields */ }

Condition variable which allows to use multiple conditions in MultiConditionVariable::blocking_wait_while() and MultiConditionVariable::timed_wait_while() concurrently but with the draw back that only all waiters can be triggered and not one. The reason is when one waits on multiple conditions but triggers only one waiter it is possible that a waiter is triggered which condition is not satisfied while another waiter is not signaled which could have continued. This could lead to some deadlocks.

The condition variable comes also with an integrated mutex to avoid undefined behavior when multiple waiters try to wait on different mutexes. With an integrated mutex this problem is solved.

The underlying data can be read/written and waiters can be triggered as soon as the data is written. The condition variable provides the following features:

wait on the condition variable: wait, timed_wait

wait until a defined condition occurs: blocking_wait_while, timed_wait_while

modify condition variable and then notify waiters: notify_all, modify_notify_all

trigger waiters without changing condition variable: trigger_all

change the underlying value without notifying any waiter: lock, try_lock, timed_lock

Example

a detailed builder example can be found here ConditionVariableBuilder

use iceoryx2_bb_posix::condition_variable::*;
use std::thread;
use std::sync::Arc;
use std::time::Duration;

let mtx_handle = MutexHandle::<i32>::new();
let cv = ConditionVariableBuilder::new()
                .create_multi_condition_variable(1234, &mtx_handle).expect("failed to create cv");

thread::scope(|s| {
    let t1 = s.spawn(|| {
        // wait until value is 5000
        let guard = cv.blocking_wait_while(|t| *t == 5000).expect("failed to wait");
        println!("cv value changed to 5000");
    });

    let t2 = s.spawn(|| {
        // wait for 1 second until value is greater 4000
        let result = cv.timed_wait_while(Duration::from_secs(1), |t| *t > 4000).expect("failed to wait");
        match result {
            None => println!("timeout reached"),
            Some(v) => println!("value {} > 4000", *v),
        }
    });

    {
        // acquire notification guard
        let mut guard = cv.notify_all().expect("failed to notify");
        // change underlying value to 4005
        *guard = 4005;
        // the guard goes out of scope and notifies all waiters
    }

    // apply clojure, set value to 4008 and then trigger waiters
    cv.modify_notify_all(|value| { *value = 4008 }).expect("failed to notify");

    {
        // acquire guard to the underlying value without triggering the condition variable
        let mut guard = cv.lock().expect("failed to acquire value");
        // read the underlying value
        println!("current value is {}", *guard);
        // set the underlying value to 5000
        *guard = 5000;
    }

    // trigger condition variable manually
    cv.trigger_all();
});

Implementations

impl<'mtx_handle, T> MultiConditionVariable<'mtx_handle, T>
where T: Debug,

pub fn blocking_wait_while<P>( &self, predicate: P, ) -> Result<MutexGuard<'_, '_, T>, ConditionVariableWaitError<'_, '_, T>>
where P: FnMut(&mut T) -> bool,

Waits until the condition variable was signaled by MultiConditionVariable::notify_all(), MultiConditionVariable::modify_notify_all() or BasicConditionVariableInterface::trigger_all() and the provided predicate returns true.

pub fn timed_wait_while<P>( &self, timeout: Duration, predicate: P, ) -> Result<Option<MutexGuard<'_, '_, T>>, ConditionVariableTimedWaitError<'_, '_, T>>
where P: FnMut(&mut T) -> bool,

Waits until the condition variable was signaled by MultiConditionVariable::notify_all(), MultiConditionVariable::modify_notify_all() or BasicConditionVariableInterface::trigger_all() and the provided predicate returns true or the provided timeout has elapsed.

pub fn notify_all( &self, ) -> Result<MultiConditionVariableGuard<'_, '_, '_, T>, MutexLockError<'_, '_, T>>

Returns a MultiConditionVariableGuard which provides read/write access to the underlying value of the condition variable, like the MutexGuard. When it goes out of scope all waiters will be notified.

pub fn modify_notify_all<Modifier>( &self, modifier: Modifier, ) -> Result<(), MutexLockError<'_, '_, T>>
where Modifier: FnOnce(&mut T),

Applies the modifier to the underlying value and notifies all waiters.