foundation::prelude
Trait BasicConditionVariableInterface

pub trait BasicConditionVariableInterface<T>:
    Sized
    + ConditionVariableHandle<T>
    + Debug
where
    T: Debug,
{
    // Provided methods
    fn wait(
        &self,
    ) -> Result<MutexGuard<'_, '_, Self::Type>, ConditionVariableWaitError<'_, '_, Self::Type>> { ... }
    fn timed_wait(
        &self,
        timeout: Duration,
    ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, ConditionVariableTimedWaitError<'_, '_, Self::Type>> { ... }
    fn trigger_all(&self) { ... }
    fn lock(
        &self,
    ) -> Result<MutexGuard<'_, '_, Self::Type>, MutexLockError<'_, '_, Self::Type>> { ... }
    fn try_lock(
        &self,
    ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, MutexLockError<'_, '_, Self::Type>> { ... }
    fn timed_lock(
        &self,
        timeout: Duration,
    ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, MutexTimedLockError<'_, '_, Self::Type>> { ... }
}

Provided Methods

fn wait( &self, ) -> Result<MutexGuard<'_, '_, Self::Type>, ConditionVariableWaitError<'_, '_, Self::Type>>

Waits until the condition variable was signaled by notify_, modify_notify_ or trigger_*.

fn timed_wait( &self, timeout: Duration, ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, ConditionVariableTimedWaitError<'_, '_, Self::Type>>

Waits until the condition variable was signaled by notify_, modify_notify_ or trigger_* or the provided timeout has elapsed.

fn trigger_all(&self)

Notifies all waiters.

fn lock( &self, ) -> Result<MutexGuard<'_, '_, Self::Type>, MutexLockError<'_, '_, Self::Type>>

Returns a MutexGuard which can be used to read/write the underlying value of the condition variable.

fn try_lock( &self, ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, MutexLockError<'_, '_, Self::Type>>

Tries to return a MutexGuard which can be used to read/write the underlying value of the condition variable. If someone else already holds the lock of the underlying mutex None is returned.

fn timed_lock( &self, timeout: Duration, ) -> Result<Option<MutexGuard<'_, '_, Self::Type>>, MutexTimedLockError<'_, '_, Self::Type>>

Tries to return a MutexGuard which can be used to read/write the underlying value of the condition variable. When the timeout has passed it returns None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.